FaceForge_Backend/requirements.txt
FastAPI==0.111.0
uvicorn==0.29.0
google-generativeai==0.7.2
huggingface-hub==0.23.2
Pillow==10.3.0
opencv-python-headless==4.9.0.80
reportlab==4.2.0
supabase==2.4.6
PyJWT==2.8.0
bcrypt==4.1.3
python-dotenv==1.0.1
numpy<2.0.0
